Proyecto subir github/dw_api/pyproject.toml
[project]
name = "apis"
version = "1.0.0"
description = "API compartida del Proyecto RA2 - Polymarket Data Warehouse"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
]